factored out firebase and moc call
